move animal extension to extensions folder
